move pest did parser to ssi-dids
