+++ Cargo.toml
@@ -36,8 +36,6 @@ zeroize = { version = "1.4", features = ["zeroize_derive"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 serde_jcs = "0.1"
-pest = "2.1"
-pest_derive = "2.1"
 derive_builder = "0.9"
 base64 = "0.12"
 ring = { version = "0.16", optional = true }
