+++ ssi-dids/Cargo.toml
@@ -3,9 +3,14 @@ name = "ssi-dids"
 version = "0.1.0"
 edition = "2021"
 
+[features]
+default = ["parser"]
+parser = ["pest", "pest_derive"]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+pest = { version = "2.1", optional = true }
+pest_derive = { version = "2.1", optional = true }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 serde_urlencoded = "0.7"
